Refactor SSH transport module into smaller components
